refactored package aw_analysis to aw_query
