+++ Makefile
@@ -10,7 +10,7 @@ build:
 	pip3 install $(pip_install_args)
 
 test:
-	python3 -m pytest tests -v --cov=aw_core --cov=aw_datastore --cov=aw_transform --cov=aw_analysis
+	python3 -m pytest tests -v --cov=aw_core --cov=aw_datastore --cov=aw_transform --cov=aw_query
 
 coverage_html: test
 	python3 -m coverage html -d coverage_html
@@ -19,10 +19,10 @@ benchmark:
 	python3 -m aw_datastore.benchmark
 
 typecheck:
-	export MYPYPATH=./stubs; python3 -m mypy aw_core aw_datastore aw_transform aw_analysis --show-traceback --ignore-missing-imports --follow-imports=skip
+	export MYPYPATH=./stubs; python3 -m mypy aw_core aw_datastore aw_transform aw_query --show-traceback --ignore-missing-imports --follow-imports=skip
 
 typecheck-strict:
-	export MYPYPATH=./stubs; python3 -m mypy aw_core aw_datastore aw_transform aw_analysis --strict-optional --check-untyped-defs; echo "Not a failing step"
+	export MYPYPATH=./stubs; python3 -m mypy aw_core aw_datastore aw_transform aw_query --strict-optional --check-untyped-defs; echo "Not a failing step"
 
 clean:
 	rm -rf build dist
