--- aw_analysis/__init__.py
@@ -1,2 +1,0 @@
-from . import query2_error
-from . import query2
