+++ aw_query/__init__.py
@@ -1,0 +1,4 @@
+from . import exceptions
+from . import query2
+
+from .query2 import query
