+++ aw_query/exceptions.py
@@ -1,11 +1,14 @@
 class QueryException(Exception):
     pass
 
+
 class QueryFunctionException(QueryException):
     pass
 
+
 class QueryParseException(QueryException):
     pass
+
 
 class QueryInterpretException(QueryException):
     pass
